Update to rust-fuse v0.3.1
This release adds socket file type support.

rust-fuse also changed the type of the `offset` parameter of the `read`,
`write`, and `readdir` calls from `u64` to `i64` to match the type used
by the kernel, but these values should never actually be negative
because that wouldn't make sense (they're absolute offsets, not
relative). fuse-mt checks and returns EINVAL if they are, and then
casts them to u64 and passes them on to the implementation. So this is
not a breaking change for fuse-mt users as it is for rust-fuse users.